Change provider to bloc for shipments
