+++ lib/app/shipments/presentation/bloc/shipment_state.dart
@@ -1,13 +1,11 @@
 part of 'shipment_bloc.dart';
 
-enum ShipmentStatus { initial, loading, success, error, selected }
+enum ShipmentStatus { initial, success, error }
 
 extension ShipmentStatusChange on ShipmentStatus {
   bool get isInitial => this == ShipmentStatus.initial;
-  bool get isLoading => this == ShipmentStatus.loading;
   bool get isSuccess => this == ShipmentStatus.success;
   bool get isError => this == ShipmentStatus.error;
-  bool get isSelected => this == ShipmentStatus.selected;
 }
 
 final class ShipmentState extends Equatable {
