+++ pubspec.yaml
@@ -46,6 +46,8 @@ dependencies:
   font_awesome_flutter: ^10.5.0
   infinite_scroll_pagination: ^4.0.0
   flutter_bloc: ^8.1.3
+  stream_transform: ^2.1.0
+  rxdart: ^0.27.7
 
 dev_dependencies:
   flutter_test:
